let Poke::Store handle serialization/deserialization
Also handle non-JSON items in db.  There shouldn't be any such in the
datastore, but in the weird case where some is found, I don't see the
point of chucking a JSON parser error.  Instead we return a faux
'text/plain' item with the contents as whatever string was actually
found.  Err on the side of returning data.