refactored leveldb data storage
It needed levelDb objects to be used in a Singleton way, and each
instance shared a lot of behavior.  So I've revised Poke::Store to be a
base class which exposes a .store= method on each class to allow for
configuration (which is now done in each spec as needed, and all
together in the Poke module for the purposes of the app).