renamed quota-related models
After extracting the token chain to a gem, it became clear that the use
of the name Token and TokenAnchorChain were somewhat confusing (as model
names).  Accordingly they are renamed to QuotaToken and QuotaAnchor,
respectively.

Initially I had thought to name QuotaAnchor simply Quota, but there is
already a middleware in the project with that name, so I couldn't.

Broader observation: because of the structure of rack/middleware, it
seems like it would be helpful to give middlewares a name which includes
a verb which indicates its function.  Following this approach would
certainly evade a collision on the "Quota" namespace.